refactor:  moved constituent parameters function from predict to arguments
feat: add functions for tide generating forces and potentials
