+++ doc/source/api_reference/predict.rst
@@ -36,8 +36,6 @@ Calling Sequence
 
 .. autofunction:: pyTMD.predict.solid_earth_tide
 
-.. autofunction:: pyTMD.predict._constituent_parameters
-
 .. autofunction:: pyTMD.predict._out_of_phase_diurnal
 
 .. autofunction:: pyTMD.predict._out_of_phase_semidiurnal
