+++ pyTMD/solve.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 solve.py
-Written by Tyler Sutterley (12/2023)
+Written by Tyler Sutterley (01/2024)
 Routines for estimating the harmonic constants for ocean tides
 
 REFERENCES:
@@ -21,6 +21,7 @@ PROGRAM DEPENDENCIES:
     predict.py: predict tidal values using harmonic constants
 
 UPDATE HISTORY:
+    Updated 01/2024: add functions for tide generating forces and potentials
     Written 12/2023
 """
 
@@ -28,8 +29,11 @@ from __future__ import annotations
 
 import numpy as np
 import scipy.linalg
-import pyTMD.predict
-from pyTMD.arguments import arguments
+import pyTMD.arguments
+from pyTMD.constants import constants
+
+# get WGS84 ellipsoid parameters in meters-kg-seconds
+_wgs84 = constants(ellipsoid='WGS84', units='MKS')
 
 def constants(t: float | np.ndarray,
         ht: np.ndarray,
@@ -88,7 +92,7 @@ def constants(t: float | np.ndarray,
 
     # load the nodal corrections
     # convert time to Modified Julian Days (MJD)
-    pu, pf, G = arguments(t + 48622.0, constituents,
+    pu, pf, G = pyTMD.arguments.arguments(t + 48622.0, constituents,
         deltat=deltat, corrections=corrections)
 
     # create design matrix
@@ -99,7 +103,7 @@ def constants(t: float | np.ndarray,
     for k,c in enumerate(constituents):
         if corrections in ('OTIS', 'ATLAS', 'TMD3', 'netcdf'):
             amp, ph, omega, alpha, species = \
-                pyTMD.predict._constituent_parameters(c)
+                pyTMD.arguments._constituent_parameters(c)
             th = omega*t*86400.0 + ph + pu[:,k]
         elif corrections in ('GOT', 'FES'):
             th = G[:,k]*np.pi/180.0 + pu[:,k]
@@ -129,3 +133,136 @@ def constants(t: float | np.ndarray,
 
     # return the amplitude and phase
     return (amp, phase)
+
+# PURPOSE: calculate the astronomical tide generating force
+def _generating_force(
+        lon: np.ndarray,
+        lat: np.ndarray,
+        c: str,
+    ):
+    """
+    Computes the astronomical tide generating force for a tidal
+    constituent [1]_
+
+    Parameters
+    ----------
+    lon: np.ndarray
+        longitudes of grid centers (degrees east)
+    lat: np.ndarray
+        latitudes of grid centers (degrees north)
+    c: str
+        tidal constituent ID
+
+    References
+    ----------
+    .. [1] G. D. Egbert and S. Y. Erofeeva, "Efficient Inverse Modeling of
+        Barotropic Ocean Tides," *Journal of Atmospheric and Oceanic
+        Technology*, 19(2), 183--204, (2002).
+        `doi: 10.1175/1520-0426(2002)019<0183:EIMOBO>2.0.CO;2`__
+
+    .. __: https://doi.org/10.1175/1520-0426(2002)019<0183:EIMOBO>2.0.CO;2
+    """
+
+    # grid spacing in radians
+    dph = (lon[1] - lon[0])*np.pi/180.0
+    dth = (lat[1] - lat[0])*np.pi/180.0
+    # calculate meshgrid from latitude and longitude
+    gridlon, gridlat = np.meshgrid(lon, lat)
+
+    # longitude (in radians) of u and v nodes
+    phi_v = gridlon*np.pi/180.0
+    phi_u = phi_u - dph/2.0
+
+    # colatitudes (in radians) of u and v nodes
+    th_u = (90.0 - gridlat)*np.pi/180.0
+    th_v = th_v + dth/2.0
+
+    # load parameters for each constituent
+    amp, ph, omega, alpha, species = \
+        pyTMD.arguments._constituent_parameters(c)
+    # uniform zonal dependence of astronomical forcing
+    ph = 0.0
+
+    # loading effects
+    c = (alpha*_wgs84.gamma*amp/_wgs84.rad_e)
+    # calculate forcing for constituent
+    Fu = c*np.exp(1j*phi_u*species + 1j*ph)
+    Fv = c*np.exp(1j*phi_v*species + 1j*ph)
+    # calculate latitudinal dependence of forcing
+    # for a given spherical harmonic dependence
+    if (species == 1):
+        # diurnal species
+        Fu *= 2j*np.cos(th_u)
+        Fv *= 2.0*(2.0*np.sin(th_v)**2 - 1.0)
+    elif (species == 2):
+        # semidiurnal species
+        Fu *= 2j*np.sin(th_u)
+        Fv *= -2.0*(np.sin(th_v)*np.cos(th_v))
+    else:
+        # long-period species
+        Fu *= 0.0 + 0j
+        Fv *= -3.0*(np.sin(th_v)*np.cos(th_v))
+    # return the generating forces
+    return (Fu, Fv)
+
+# PURPOSE: calculate the astronomical tide generating potential
+# converted to an equilibrium tide elevation
+def _generating_potential(
+        lon: np.ndarray,
+        lat: np.ndarray,
+        c: str,
+    ):
+    """
+    Computes the astronomical tide generating potential for a tidal
+    constituent converted to an equilibrium tide elevation [1]_
+
+    Parameters
+    ----------
+    lon: np.ndarray
+        longitudes of grid centers (degrees east)
+    lat: np.ndarray
+        latitudes of grid centers (degrees north)
+    c: str
+        tidal constituent ID
+
+    References
+    ----------
+    .. [1] G. D. Egbert and S. Y. Erofeeva, "Efficient Inverse Modeling of
+        Barotropic Ocean Tides," *Journal of Atmospheric and Oceanic
+        Technology*, 19(2), 183--204, (2002).
+        `doi: 10.1175/1520-0426(2002)019<0183:EIMOBO>2.0.CO;2`__
+
+    .. __: https://doi.org/10.1175/1520-0426(2002)019<0183:EIMOBO>2.0.CO;2
+    """
+
+    # calculate meshgrid from latitude and longitude
+    gridlon, gridlat = np.meshgrid(lon, lat)
+
+    # longitude (in radians)
+    phi_h = gridlon*np.pi/180.0
+    # colatitudes (in radians)
+    th_h = (90.0 - gridlat)*np.pi/180.0
+
+    # load parameters for each constituent
+    amp, ph, omega, alpha, species = \
+        pyTMD.arguments._constituent_parameters(c)
+    # uniform zonal dependence of astronomical potential
+    ph = 0.0
+
+    # calculate potential for constituent
+    c = (alpha*amp)
+    zeta = c*np.exp(1j*phi_h*species + 1j*ph)
+
+    # calculate latitudinal dependence of potential
+    # for a given spherical harmonic dependence
+    if (species == 1):
+        # diurnal species
+        zeta *= 2*0*np.sin(th_h)*np.cos(th_h)
+    elif (species == 2):
+        # semidiurnal species
+        zeta *= np.sin(th_h)**2
+    else:
+        # long-period species
+        zeta *= -1.0*(3.0/2.0*np.cos(th_h)**2 - 1.0/2.0)
+    # return the generating potential
+    return zeta
